Rename readFile to parse_csv.py
